Feat: 모바일 종류별 헤더 분리
