+++ src/components/Layout/Header/Default/DefaultHeader.module.scss
@@ -1,0 +1,242 @@
+@use "@/styles/globals.scss" as *;
+
+.cursor {
+  cursor: pointer;
+}
+
+.container {
+  padding: 10px 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  @include Size("mobile") {
+    padding: 12px 16px;
+    max-width: 100%;
+  }
+}
+
+.leftSection {
+  display: flex;
+  align-items: center;
+  gap: 50px;
+}
+
+.navUpload {
+  display: flex;
+  flex-direction: column;
+}
+
+.nav {
+  display: flex;
+  align-items: center;
+  position: relative;
+
+  @include Size("mobile") {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 20px;
+  }
+
+  @include Size("tablet") {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 20px;
+  }
+
+  .navItem {
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+    height: 25px;
+
+    .item {
+      color: $gray60;
+      @include Sub2;
+      display: inline-block;
+      margin: 0;
+      transition: color 0.3s;
+
+      &.active {
+        color: $gray70;
+      }
+
+      @include Size("mobile") {
+        padding: 10px 0;
+        @include Sub2;
+        font-size: 18px;
+      }
+
+      @include Size("tablet") {
+        padding: 14px 0;
+        @include Title3;
+      }
+    }
+  }
+
+  .indicator {
+    position: absolute;
+    height: 2px;
+    background-color: $gray80;
+    transition: transform 0.3s ease, width 0.3s ease;
+    bottom: -8px;
+  }
+}
+
+.wrapper {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+  flex-shrink: 0;
+
+  @include Size("mobile") {
+    gap: 0;
+  }
+}
+
+.profileContact {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  position: relative;
+}
+
+.profileSection {
+  display: flex;
+  align-items: center;
+  gap: 24px;
+  position: relative;
+}
+
+.profileImage {
+  width: 28px;
+  height: 28px;
+  object-fit: cover;
+  border-radius: 50%;
+  overflow: hidden;
+  cursor: pointer;
+
+  @include Size("mobile") {
+    width: 40px;
+    height: 40px;
+  }
+}
+
+.dropdown {
+  position: absolute;
+  padding: 8px 0;
+  top: 40px;
+  right: 0;
+  background: $gray0;
+  box-shadow: 0px 0px 12px 0px rgba(0, 0, 0, 0.15);
+  border-radius: 12px;
+  z-index: 10;
+  width: 150px;
+  overflow: hidden;
+
+  .dropdownItem {
+    display: flex;
+    align-items: center;
+    padding: 10px 16px;
+    cursor: pointer;
+    @include Sub4;
+    color: $gray80;
+
+    &:hover {
+      background: $gray10;
+    }
+
+    &.logout {
+      color: $gray60;
+      @include Sub4;
+    }
+
+    .dropdownProfileImage {
+      width: 28px;
+      height: 28px;
+      border: 1px solid $gray20;
+      object-fit: cover;
+      border-radius: 50px;
+      overflow: hidden;
+    }
+
+    span {
+      padding-left: 8px;
+    }
+  }
+
+  .divider {
+    height: 1px;
+    background: #e0e0e0;
+    margin: 4px 0;
+  }
+}
+
+.icons {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+
+  @include Size("mobile") {
+    gap: 0;
+    margin-left: 8px;
+  }
+}
+
+.uploadBtn {
+  &:hover {
+    animation: pulse 0.6s ease-in-out;
+  }
+}
+
+.searchbarContainer {
+  width: 290px;
+  height: 40px;
+  padding-left: 16px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+  border-radius: 12px;
+  border: 1px solid $gray30;
+  background-color: $gray10;
+  transition: transform 0.3s ease, width 0.3s ease;
+
+  .input {
+    width: 100%;
+    color: $gray70;
+    @include Label2;
+    background: none;
+    border: none;
+    outline: none;
+
+    &::placeholder {
+      color: $gray50;
+    }
+  }
+}
+
+.notificationWrapper {
+  position: relative;
+
+  .notification {
+    cursor: pointer;
+    position: relative;
+  }
+}
+
+.menuContentWrapper {
+  flex-grow: 1;
+}
+
+.mobileFooterWrap {
+  width: 100%;
+  flex-shrink: 0;
+  position: relative;
+  height: auto;
+  flex-direction: column-reverse;
+  align-items: flex-start;
+  margin-top: 24px;
+}
