+++ src/components/Layout/Header/Popup/PopupHeader.module.scss
@@ -1,0 +1,30 @@
+@use "@/styles/globals.scss" as *;
+
+.container {
+  padding: 10px 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: relative;
+
+  @include Size("mobile") {
+    padding: 12px 16px;
+    max-width: 100%;
+  }
+}
+
+.closeButton {
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 8px;
+}
+
+.title {
+  @include Title2;
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+}
