+++ src/components/Layout/Header/Popup/PopupHeader.types.ts
@@ -1,0 +1,4 @@
+export interface PopupHeaderProps {
+  onClose: () => void;
+  title: string;
+}
