+++ src/components/Layout/Header/Sub/SubHeader.module.scss
@@ -1,0 +1,32 @@
+@use "@/styles/globals.scss" as *;
+
+.container {
+  padding: 10px 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  @include Size("mobile") {
+    padding: 12px 16px;
+    max-width: 100%;
+  }
+}
+
+.backButton {
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 8px;
+}
+
+.icons {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+
+  @include Size("mobile") {
+    gap: 0;
+  }
+}
