+++ src/components/Layout/Header/types/Header.types.ts
@@ -1,0 +1,3 @@
+export interface HeaderProps {
+  variant?: "default" | "sub";
+}
