+++ src/components/Layout/SideMenu/SideMenu.module.scss
@@ -1,0 +1,225 @@
+@use "@/styles/globals.scss" as *;
+
+.overlay {
+  background-color: rgba(0, 0, 0, 0.15);
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  opacity: 0;
+  pointer-events: none;
+  transition: opacity 0.3s ease-in-out;
+  z-index: 1500;
+
+  &.open {
+    opacity: 1;
+    pointer-events: auto;
+  }
+
+  .uploadBtnContainer {
+    margin-top: 16px;
+    padding-bottom: 30px;
+  }
+
+  .sideMenu {
+    position: fixed;
+    top: 0;
+    right: 0;
+    width: 240px;
+    height: 100vh;
+    padding: 16px 20px;
+    background-color: $gray0;
+    transform: translateX(100%);
+    transition: transform 0.3s ease-in-out;
+    z-index: 1600;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+    @include Size("mobile") {
+      width: 240px;
+      height: 100%;
+      padding: 16px 20px;
+    }
+
+    &.open {
+      transform: translateX(0);
+    }
+
+    .closeBtn {
+      display: flex;
+      justify-content: flex-end;
+    }
+
+    .bar {
+      width: 100%;
+      height: 1px;
+      background-color: $gray30;
+      margin-top: 4px;
+      margin-bottom: 24px;
+    }
+
+    .mobileProfile {
+      display: flex;
+      flex-direction: column;
+
+      .mobileMyInfo {
+        width: 100%;
+        display: flex;
+        align-items: center;
+        gap: 8px;
+      }
+
+      .profileSubmenu {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        margin-top: 16px;
+        margin-bottom: 20px;
+
+        &.profileSubmenuOpen {
+          margin-bottom: 10px;
+        }
+      }
+
+      .name {
+        @include Sub4;
+        color: $gray80;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        display: -webkit-box;
+        -webkit-line-clamp: 1;
+        line-clamp: 1;
+        -webkit-box-orient: vertical;
+      }
+
+      .submenu {
+        display: flex;
+        width: 32px;
+        height: 32px;
+        padding: 6px;
+        justify-content: center;
+        align-items: center;
+        flex-shrink: 0;
+        border-radius: 8px;
+        border: 1px solid $gray30;
+        background: $gray0;
+      }
+
+      .btns {
+        display: flex;
+        align-items: flex-start;
+        flex-direction: column;
+        gap: 2px;
+        margin-bottom: 20px;
+      }
+
+      .itemBtn {
+        @include Label1;
+        color: $gray60;
+        display: flex;
+        align-items: center;
+        gap: 10px;
+        background: none;
+        height: 34px;
+        border: none;
+        padding: 10px 0;
+      }
+
+      .uploadFeedBtn {
+        margin-bottom: 20px;
+      }
+
+      .mobileDropdown {
+        cursor: pointer;
+        position: relative;
+      }
+
+      .menuContent {
+        display: flex;
+        flex-direction: column;
+      }
+    }
+  }
+}
+.menuContentWrapper {
+  flex-grow: 1;
+}
+
+.mobileFooterWrap {
+  width: 100%;
+  flex-shrink: 0;
+  position: relative;
+  height: auto;
+  flex-direction: column-reverse;
+  align-items: flex-start;
+  margin-top: 24px;
+}
+
+.nav {
+  display: flex;
+  align-items: center;
+  position: relative;
+
+  @include Size("mobile") {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 20px;
+  }
+
+  @include Size("tablet") {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 20px;
+  }
+
+  .navItem {
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+    height: 25px;
+
+    .item {
+      color: $gray60;
+      @include Sub2;
+      display: inline-block;
+      margin: 0;
+      transition: color 0.3s;
+
+      &.active {
+        color: $gray70;
+      }
+
+      @include Size("mobile") {
+        padding: 10px 0;
+        @include Sub2;
+        font-size: 18px;
+      }
+
+      @include Size("tablet") {
+        padding: 14px 0;
+        @include Title3;
+      }
+    }
+  }
+}
+
+.profileImage {
+  width: 28px;
+  height: 28px;
+  object-fit: cover;
+  border-radius: 50%;
+  overflow: hidden;
+  cursor: pointer;
+
+  @include Size("mobile") {
+    width: 40px;
+    height: 40px;
+  }
+}
+
+.navUpload {
+  display: flex;
+  flex-direction: column;
+}
